# docs/openapi/agent.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jpcite Agent Evidence API",
    "description": "Agent-safe OpenAPI subset for evidence prefetch before answer generation. It excludes billing, webhook, OAuth, account-management, and operator endpoints. Anonymous callers can evaluate within the published daily limit unless an operation marks X-API-Key as required; callers that need higher volume send X-API-Key.",
    "version": "0.3.2",
    "contact": {
      "name": "jpcite Support",
      "url": "https://jpcite.com/tokushoho.html"
    },
    "termsOfService": "https://jpcite.com/tos.html",
    "license": {
      "name": "Proprietary - see termsOfService"
    }
  },
  "paths": {
...
    "/v1/programs/batch": {
      "post": {
        "tags": [
          "programs"
        ],
        "summary": "Batch fetch up to 50 programs by unified_id",
        "description": "Resolve up to 50 `unified_id` values in a single round-trip. Output shape matches `GET /v1/programs/{unified_id}` per record, so SDK callers can `chunk(ids, 50)` and stitch locally without per-id round trips. The 50-cap *is* the pagination — there is no page envelope.\n\n**Order contract:** `results[i]` corresponds to the i-th deduped input id (first occurrence wins). Missing ids go to `not_found[]` — this is NOT a 404, partial success is the point. Batch hard-codes `fields=full` so anonymous tier callers must upgrade (sequential GETs with `fields=default` remain anonymous-OK).",
        "operationId": "batch_get_programs_v1_programs_batch_post",
        "parameters": [
          {
            "name": "X-Cost-Cap-JPY",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cost-Cap-Jpy"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchGetProgramsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch ProgramDetail lookup. `results[]` is ordered by the deduped input `unified_ids`. Ids not found in the DB go to `not_found` — this is NOT a 404, because partial success is the point of batch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchGetProgramsResponse"
                },
                "example": {
                  "results": [
                    {
                      "unified_id": "UNI-2611050f9a",
                      "primary_name": "小規模事業者持続化補助金",
                      "tier": "B",
                      "authority_level": "national",
                      "authority_name": "日本商工会議所/全国商工会連合会",
                      "program_kind": "subsidy",
                      "amount_max_man_yen": 200.0,
                      "official_url": "https://r3.jizokukahojokin.info/",
                      "source_fetched_at": "2026-04-22T13:20:57Z"
                    },
                    {
                      "unified_id": "UNI-185c08e0c1",
                      "primary_name": "デジタル化・AI導入補助金（旧IT導入補助金）",
                      "tier": "B",
                      "authority_level": "national",
                      "authority_name": "国（農水省等）",
                      "program_kind": "subsidy",
                      "amount_max_man_yen": 450.0,
                      "official_url": "https://it-shien.smrj.go.jp/",
                      "source_fetched_at": "2026-04-22T13:20:57Z"
                    }
                  ],
                  "not_found": [
                    "UNI-deadbeef00"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error — `code` ∈ {invalid_enum, invalid_date_format, missing_required_arg, out_of_range, ambiguous_query}.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid authentication — returned only when an API key is supplied but invalid. Anonymous callers may use the published daily limit; quota exhaustion returns 429.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found — `code` ∈ {no_matching_records, seed_not_found, route_not_found}.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "input validation failed (empty list, >50 ids, bad shape). `error.code='invalid_enum'`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit — `code='rate_limit_exceeded'`. Honour `Retry-After`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Server error — `code` ∈ {service, internal_error, db_locked, db_unavailable}.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Subsystem unavailable — `code` ∈ {subsystem_unavailable, service_unavailable, cap_reached}.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {}
        ],
        "x-jpcite-agent-safe": true,
        "x-jpcite-auth": "optional_x_api_key_for_paid_volume"
      }
    },
...
    "/v1/cost/preview": {
      "post": {
        "tags": [
          "cost"
        ],
        "summary": "Predict the cost of a planned tool-call stack (no request charge, ¥0)",
        "description": "Returns a no-charge cost estimate for planned jpcite tool calls without executing them.\n\n**Pricing**: metered calls are ¥3/req (税別; 税込 ¥3.30). The preview itself is not metered and is rate-limited 50/min per IP/key.\n\n**Tax-related previews**: when the stack touches tax-relevant tools (`evaluate_tax_ruleset`, `search_tax_incentives`, etc.) the response includes a disclaimer for downstream display.",
        "operationId": "post_cost_preview_v1_cost_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {}
        ],
        "x-jpcite-agent-safe": true,
        "x-jpcite-auth": "optional_x_api_key_for_paid_volume"
      }
    }
  },
  "components": {
    "schemas": {
      "BatchGetProgramsRequest": {
        "properties": {
          "unified_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Unified Ids"
          },
          "max_cost_jpy": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cost Jpy",
            "description": "Optional body cost cap. For paid batch requests, this or `X-Cost-Cap-JPY` is required before billing; the lower cap binds."
          }
        },
        "type": "object",
        "required": [
          "unified_ids"
        ],
        "title": "BatchGetProgramsRequest",
        "description": "Body for POST /v1/programs/batch.\n\n`unified_ids` is capped at 50. Submit up to 50 unified_ids per request;\nlarger lists should be split client-side. Exceeding the cap returns HTTP 422."
      },
      "BatchGetProgramsResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ProgramDetail"
            },
            "type": "array",
            "title": "Results"
          },
          "not_found": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Not Found"
          }
        },
        "type": "object",
        "required": [
          "results",
          "not_found"
        ],
        "title": "BatchGetProgramsResponse",
        "description": "Response for POST /v1/programs/batch.\n\n`results[]` is in the same order as the deduped input `unified_ids`.\nMissing ids go to `not_found` (NOT a 404 — partial success is the point\nof batch-fetch). Use `not_found` length == 0 as the \"everything resolved\"\nsignal."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Customer API key issued via Stripe Checkout. Anonymous tier (no key) gets 3 req/日 per IP."
      }
    }
  },
  "servers": [
    {
      "url": "https://api.jpcite.com",
      "description": "Production"
    }
  ],
  "security": []
}